requirements.txt
# TCA RCA Agent - Python Dependencies

# Claude AI & Agent SDK
anthropic>=0.40.0
claude-agent-sdk>=1.0.0

# FastAPI & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
sse-starlette>=1.6.5

# Memory & Learning
mem0ai>=0.1.0

# GitHub Integration
PyGithub>=2.1.1
gitpython>=3.1.40

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Code Analysis
ast-grep-py>=0.1.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=23.11.0
isort>=5.12.0
ruff>=0.1.6
mypy>=1.7.0
